requirements.txt
argparse
langchain
langchain_openai
youtube_transcript_api
python-dotenv